Refactoring genetics into genetics-tutor
